dagster_pipeline/pyproject.toml
[project]
name = "dagster_acled"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-duckdb>=0.26.20",
    "dagster-webserver>=1.10.20",
    "aiohttp>=3.12.13",
    "aws-advanced-python-wrapper>=1.2.0",
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.9.0",
    "boto3>=1.39.1",
    "dagster-duckdb>=0.27.1",
    "dagster-webserver>=1.10.20",
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "polars>=1.31.0",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
    "xlsxwriter>=3.2.5",
    "dagster-aws>=0.27.2",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dagster_acled.definitions"
project_name = "dagster_acled"

[tool.setuptools.packages.find]
exclude=["dagster_acled_tests"]
